Reorganize plugin src directory tree
