+++ server/src/plugins/metadata.rs
@@ -2,7 +2,8 @@ use std::{path::PathBuf, collections::BTreeMap};
 use serde::{Serialize, Deserialize};
 use uuid::Uuid;
 
-use super::executors::{PluginManager, DnsPlugin};
+use super::manager::PluginManager;
+use super::plugin::DnsPlugin;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct PluginMetadata {
@@ -47,7 +48,7 @@ impl DnsPlugin {
             home: self.home_dir().to_owned(),
             is_listener: self.is_listener(),
             is_interceptor: false, // Not implemented
-            is_resolver: self.is_resolver(), // Not implemented
+            is_resolver: self.is_resolver(),
             is_validator: false, // Not implemented
             is_inspector: false, // Not implemented
             intercept_position: None,
