+++ server/src/plugins/mod.rs
@@ -1,6 +1,6 @@
 
-pub mod executors;
-pub mod loaders;
+pub mod manager;
+pub mod plugin;
 pub mod metadata;
 
 type ListenerDecodeFn = unsafe extern "C" fn(
