+++ server/src/plugins/plugin.rs
@@ -1,11 +1,25 @@
 
-use std::{ffi::c_char, collections::BTreeMap, sync::{Arc, Weak}, path::{PathBuf, Path}, error::Error, fmt::Display};
-
+use super::{ListenerDecodeFn, ListenerEncodeFn, ResolverFn};
 use libloading::{Symbol, Library};
 use modns_sdk::ffi;
-use uuid::Uuid;
 
-use super::{ResolverFn, ListenerEncodeFn, ListenerDecodeFn, loaders::PluginLoaderError};
+use std::error::Error;
+use std::fmt::Display;
+use std::path::{PathBuf, Path};
+use std::ffi::{OsStr, c_char};
+
+#[derive(Debug)]
+pub enum PluginLoaderError {
+    LibraryLoadError(libloading::Error),
+    NoListeners,
+    NoResolvers,
+}
+
+impl From<libloading::Error> for PluginLoaderError {
+    fn from(value: libloading::Error) -> Self {
+        Self::LibraryLoadError(value)
+    }
+}
 
 #[derive(Debug, PartialEq, Eq)]
 pub enum PluginExecutorError {
@@ -137,103 +151,38 @@ impl DnsPlugin {
     pub fn home_dir(&self) -> &Path {
         &self.home_dir
     }
+
+    pub fn load<P: AsRef<OsStr>>(home_dir: P, enable: bool) -> Result<Self, libloading::Error> {
+
+        let home_dir = PathBuf::from(home_dir.as_ref());
+
+        let lib = unsafe { Library::new(home_dir.join("plugin.so")) }?;
+
+        let is_listener =
+        get_sym::<ListenerDecodeFn>(&lib, b"impl_decode_req")?.is_some() &&
+        get_sym::<ListenerEncodeFn>(&lib, b"impl_encode_resp")?.is_some();
+
+        let is_resolver =
+        get_sym::<ResolverFn>(&lib, b"impl_resolve_req")?.is_some();
+
+        Ok(Self::new(
+            lib, 
+            is_listener,
+            is_resolver,
+            home_dir,
+            enable))
+    }
+
 }
 
-pub struct PluginManager {
-    plugins: BTreeMap<uuid::Uuid, Arc<DnsPlugin>>,
-    listener: Weak<DnsPlugin>,
-    resolver: Weak<DnsPlugin>
-}
+fn get_sym<'lib, T> (lib: &'lib Library, name: &[u8]) -> Result<Option<Symbol<'lib, T>>, libloading::Error> {
+    let sym = unsafe {
+        lib.get(name)
+    };
 
-impl PluginManager {
-    pub fn new() -> Self {
-        Self {
-            plugins: BTreeMap::new(),
-            listener: Weak::new(),
-            resolver: Weak::new()
-        }
-    }
-
-    pub fn load<P: AsRef<Path>>(&mut self, dir_path: P, enable: bool) -> Result<(), PluginLoaderError>{
-        let id = Uuid::new_v4();
-
-        let dir = PathBuf::from(dir_path.as_ref());
-
-        self.plugins.insert(
-            id,
-            Arc::new(
-                DnsPlugin::load(dir.clone(), enable)?
-            )
-        );
-
-        log::info!("Loaded plugin {id} from directory {}", dir.display());
-
-        Ok(())
-    }
-
-    pub fn search(&mut self, search_path: &[PathBuf]) {
-        for parent in search_path {
-
-            match parent.read_dir() {
-                Ok(dir_info) => {
-                    let dirs = dir_info.filter_map(|d| {
-                        match d {
-                            Ok(f) if f.file_type().ok()?.is_dir()=> Some(f.path()),
-                            Ok(_) => None,
-                            Err(_) => None,
-                        }
-                    });
-
-                    for subdir in dirs {
-                        let so_path = subdir.join("plugin.so");
-
-                        if !so_path.is_file() { continue; }
-
-                        if let Err(e) = self.load(subdir, false) {
-                            log::error!("Failed to load library at {}: {e:?}", so_path.display())
-                        }
-                    }
-                },
-                Err(e) => {
-                    log::info!("Failed to read directory {}: {}", parent.display(), e);
-                },
-            }
-        }
-
-        self.init()
-    }
-
-    pub fn init(&mut self) {
-        self.listener = match self.plugins.values().find(|p| p.is_listener()) {
-            Some(p) => Arc::downgrade(p),
-            None => Weak::new(),
-        };
-
-        self.resolver = match self.plugins.values().find(|p| p.is_resolver()) {
-            Some(p) => Arc::downgrade(p),
-            None => Weak::new(),
-        };
-    }
-
-    pub fn decode(&self, req: &[u8]) -> Result<Box<ffi::DnsMessage>, PluginExecutorError> {
-        self.listener.upgrade()
-        .ok_or(PluginExecutorError::NoneEnabled)?
-        .decode(req)
-    }
-
-    pub fn encode(&self, message: Box<ffi::DnsMessage>) -> Result<Vec<i8>, PluginExecutorError> {
-        self.listener.upgrade()
-        .ok_or(PluginExecutorError::NoneEnabled)?
-        .encode(message)
-    }
-
-    pub fn resolve(&self, request: Box<ffi::DnsMessage>) -> Result<Box<ffi::DnsMessage>, PluginExecutorError> {
-        self.resolver.upgrade()
-        .ok_or(PluginExecutorError::NoneEnabled)?
-        .resolve(request)
-    }
-
-    pub fn plugins(&self) -> &BTreeMap<uuid::Uuid, Arc<DnsPlugin>> {
-        &self.plugins
+    match sym {
+        Ok(s) => Ok(Some(s)),
+        Err(libloading::Error::DlSym{ desc: _ }) => Ok(None),
+        Err(e) => Err(e),
     }
 }
